Fix an issue with Websocket connection on client side Clean and reorganise code
