+++ server/src/App.ts
@@ -6,7 +6,6 @@ import morgan from 'morgan';
 import { Container } from 'typedi';
 import { createConnection, useContainer } from 'typeorm';
 
-import { ConnectType, WorkDoneType, WorkNewType } from './events';
 import Server from './server';
 
 useContainer(Container);
@@ -17,20 +16,6 @@ createConnection().then(async connection => {
   server.app.use(compression());
   server.app.use(helmet());
   server.app.use(morgan('short'));
-
-  server.app.get('/', (req, res) => {
-    res.send(server.wsHandler.getConnected());
-  });
-
-  server.wsHandler.on(ConnectType, (socket) => {
-    console.error('this is a simple test!');
-
-    socket.emit(WorkNewType, 'get a job!');
-
-    socket.on(WorkDoneType, (socket) =>
-      console.log('WORK IS FINISHED!')
-    );
-  });
 
   server.listen(process.env.PORT || 3000, '127.0.0.1');
